EMOTION ANALYSIS PROJECT WITH IMDB MOVIE REVIEWS
In our project, in this deep learning model, our necessary data was taken with our IMDB data set. We pulled movie genres and reviews from the dataset. According to these data, we created a learning model by making positive and negative sentiment analyzes depending on the status of the scores given by the people in the IMDB reviews. We used one of the CNN, RNN or LSTM models according to the dependence of the situations.